Module submit_solution::mock

Mock file for staking fuzzing.

Structs

ExistentialDeposit

GenesisConfig

MaxIterations

MaxNominatorRewardedPerValidator

MinimumPeriod

Origin

PalletInfo

Provides an implementation of PalletInfo to provide information about the pallet setup in the runtime.

RewardCurve

SessionKeys

Test

TestSessionHandler

Enums

Call

Event

OriginCaller

Constants

I_NPOS

Traits

InherentDataExt

Type Definitions

AccountId

AccountIndex

AllModules

All pallets included in the runtime as a nested tuple of types. Excludes the System pallet.

AllModulesWithSystem

All pallets included in the runtime as a nested tuple of types.

Balance

Balances

BalancesConfig

Block

BlockNumber

Extrinsic

Indices

IndicesConfig

Session

SessionConfig

Staking

StakingConfig

System

SystemConfig

UncheckedExtrinsic